+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.294 ] - 2026/09/11
+
+  * Updated Spotify Connect CastApp logic to better account for Cast Group leader changes.
+
 ###### [ 1.0.293 ] - 2026/09/10
 
   * Updated Spotify Connect CastApp logic to better account for Cast Group leader changes.  It now does a query immediately before starting the Spotify CastApp to determine where to send the request.